debug

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mwubneh <mwubneh@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/14 14:22:24 by mwubneh           #+#    #+#             */
-/*   Updated: 2023/01/17 14:32:03 by mwubneh          ###   ########lyon.fr   */
+/*   Updated: 2023/01/17 14:35:55 by mwubneh          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,6 +23,6 @@ int			ft_print_point(va_list arg);
 size_t		ft_print_str(const char *str);
 int			ft_print_char(char c);
 int			ft_printf(const char *str, ...);
-size_t		ft_putnbr_hexa(uint64_t nbr, char *base);
+size_t		ft_putnbr_hexa(uintptr_t nbr, char *base);
 
 #endif
